+++ src/services/Subscription/Subscription.service.js
@@ -40,6 +40,7 @@ export class Subscription {
    * @property {object} [couponCode]
    * @property {object} [giftRecipient]
    * @property {number} [subscriptionIdToRenew]
+   * @property {number} [quantity]
    */
 
   /**
@@ -145,10 +146,17 @@ export class StripeGateway {
    * @return {void}
    */
   #createSubscription = (options, callback) => {
-    const { token, plan, couponCode, product } = options;
+    const {
+      token,
+      plan,
+      couponCode,
+      product,
+      quantity = 1
+    } = options;
 
     window.Pelcro.subscription.create(
       {
+        quantity,
         gateway_token: token,
         payment_gateway: this.#paymentGateway,
         auth_token: window.Pelcro.user.read().auth_token,
@@ -176,11 +184,13 @@ export class StripeGateway {
       plan,
       couponCode,
       product,
-      giftRecipient
+      giftRecipient,
+      quantity = 1
     } = options;
 
     window.Pelcro.subscription.create(
       {
+        quantity,
         gateway_token: token,
         payment_gateway: this.#paymentGateway,
         auth_token: window.Pelcro.user.read().auth_token,
@@ -295,10 +305,17 @@ export class PaypalGateWay {
    * @return {void}
    */
   #createSubscription = (options, callback) => {
-    const { token, plan, couponCode, product } = options;
+    const {
+      token,
+      plan,
+      couponCode,
+      product,
+      quantity = 1
+    } = options;
 
     window.Pelcro.subscription.create(
       {
+        quantity,
         gateway_token: token,
         payment_gateway: this.#paymentGateway,
         auth_token: window.Pelcro.user.read().auth_token,
@@ -326,11 +343,13 @@ export class PaypalGateWay {
       plan,
       couponCode,
       product,
-      giftRecipient
+      giftRecipient,
+      quantity = 1
     } = options;
 
     window.Pelcro.subscription.create(
       {
+        quantity,
         gateway_token: token,
         payment_gateway: this.#paymentGateway,
         auth_token: window.Pelcro.user.read().auth_token,